feat: add markers for toilet positions

+++ app/src/main/java/edu/ap/WcApp/ToiletViewModel.kt
@@ -1,4 +1,4 @@
 package edu.ap.WcApp
 
-data class ToiletViewModel(val omschrijving: String, val addres: String, val lat: Float, val lon: Float, val doelgroep: String, val luiertafel: String) {
+data class ToiletViewModel(val omschrijving: String, val addres: String, val lat: Double, val lon: Double, val doelgroep: String, val luiertafel: String) {
 }
